bubbleFrame refactorisation
